Add files via upload
Here are the forecast plots of unemployment for h = 2, 12, and 24 months ahead from the ARIMA(1, 2, 1) model on the seasonally adjusted data with no predictors, in case we decide to go with this model. I generated the plots with sarima( ) and then with Arima( ). Personally, I like the Arima( ) ones better (smooth blue lines and purple confidence bands).